done c

+++ Контест 8/c.cpp
@@ -6,7 +6,7 @@
 const int64_t cInf = 2'000'000'000;
 
 struct Path {
-  size_t destination = cInf;
+  int64_t destination = cInf;
   int64_t price = cInf;
   int64_t time = cInf;
 };
@@ -14,8 +14,8 @@ struct Path {
 struct Node {
   int64_t time = cInf;
   int64_t price = cInf;
-  size_t index = cInf;
-  size_t prev_index = cInf;
+  int64_t index = cInf;
+  int64_t prev_index = cInf;
   int64_t prev_path_time = cInf;
 };
 
@@ -30,14 +30,14 @@ bool NodeCmpByPrice(const Node& left, const Node& right) {
   return left.index < right.index;
 }
 
-std::pair<int64_t, std::vector<size_t>> GetMinPath(
+std::pair<int64_t, std::vector<int64_t>> GetMinPath(
     const std::vector<std::set<Node, decltype(NodeCmpByTime)*>>& nodes) {
   auto cur_node = --(--((nodes.back()).end()));
   int64_t min_price = cur_node->price;
   if (min_price == cInf) {
     return {-1, {}};
   }
-  std::vector<size_t> ans_path(0);
+  std::vector<int64_t> ans_path(0);
   while (cur_node->index != 0) {
     ans_path.push_back(cur_node->index);
     cur_node =
@@ -52,7 +52,7 @@ auto Dijkstra(const std::vector<std::vector<Path>>& graph, int t) {
   std::vector<std::set<Node, decltype(NodeCmpByTime)*>> nodes(
       graph.size(), std::set<Node, decltype(NodeCmpByTime)*>(NodeCmpByTime));
   nodes[0].insert({0, 0, 0});
-  for (size_t i = 1; i < graph.size(); ++i) {
+  for (int64_t i = 1; i < static_cast<int64_t>(graph.size()); ++i) {
     nodes[i].insert({0, cInf, i});
     nodes[i].insert({cInf, cInf, i});
   }
@@ -84,19 +84,19 @@ auto Dijkstra(const std::vector<std::vector<Path>>& graph, int t) {
 }
 
 int main() {
-  int n;
-  int m;
-  int t;
+  int64_t n;
+  int64_t m;
+  int64_t t;
   std::cin >> n >> m >> t;
   std::vector<std::vector<Path>> graph(n);
   for (int i = 0; i < m; ++i) {
-    int start;
-    int end;
-    int price;
-    int time;
+    int64_t start;
+    int64_t end;
+    int64_t price;
+    int64_t time;
     std::cin >> start >> end >> price >> time;
-    graph[start - 1].push_back({static_cast<size_t>(end - 1), price, time});
-    graph[end - 1].push_back({static_cast<size_t>(start - 1), price, time});
+    graph[start - 1].push_back({end - 1, price, time});
+    graph[end - 1].push_back({start - 1, price, time});
   }
   auto res = Dijkstra(graph, t);
   std::cout << res.first << "\n";
